requirements.txt
ollama
faiss-cpu
tiktoken
streamlit
numpy
sentence_transformers
docling